Rename s/synchronization/sync
