+++ apollo-router/src/plugins/connectors/make_requests.rs
@@ -24,7 +24,7 @@ use crate::json_ext::PathElement;
 use crate::plugins::connectors::plugin::debug::ConnectorContext;
 use crate::services::connect;
 use crate::services::connector::request_service::Request;
-use crate::synchronization::Mutex;
+use crate::sync::Mutex;
 
 const REPRESENTATIONS_VAR: &str = "representations";
 const ENTITIES: &str = "_entities";
