+++ apollo-router/src/plugins/record_replay/replay.rs
@@ -27,7 +27,7 @@ use crate::services::execution;
 use crate::services::router;
 use crate::services::subgraph;
 use crate::services::supergraph;
-use crate::synchronization::Mutex;
+use crate::sync::Mutex;
 
 #[derive(Debug)]
 pub(crate) struct Replay {
