+++ apollo-router/src/plugins/rhai/tests.rs
@@ -43,7 +43,7 @@ use crate::services::ExecutionRequest;
 use crate::services::SubgraphRequest;
 use crate::services::SupergraphRequest;
 use crate::services::SupergraphResponse;
-use crate::synchronization::Mutex;
+use crate::sync::Mutex;
 
 // There is a lot of repetition in these tests, so I've tried to reduce that with these two
 // functions. The repetition could probably be reduced further, but ...
