+++ apollo-router/src/plugins/telemetry/config_new/connector/events.rs
@@ -19,7 +19,7 @@ use crate::plugins::telemetry::config_new::events::StandardEventConfig;
 use crate::plugins::telemetry::config_new::events::log_event;
 use crate::plugins::telemetry::config_new::extendable::Extendable;
 use crate::plugins::telemetry::config_new::instruments::Instrumented;
-use crate::synchronization::Mutex;
+use crate::sync::Mutex;
 
 #[derive(Clone)]
 pub(crate) struct ConnectorEventRequest(pub(crate) StandardEvent<ConnectorSelector>);
