+++ apollo-router/src/services/mod.rs
@@ -26,7 +26,7 @@ pub(crate) use crate::services::subgraph::Response as SubgraphResponse;
 pub(crate) use crate::services::supergraph::Request as SupergraphRequest;
 pub(crate) use crate::services::supergraph::Response as SupergraphResponse;
 pub(crate) use crate::services::supergraph::service::SupergraphCreator;
-use crate::synchronization::Mutex;
+use crate::sync::Mutex;
 
 pub(crate) mod connect;
 pub(crate) mod connector;
