+++ apollo-router/src/services/router/tests.rs
@@ -35,7 +35,7 @@ use crate::services::router::service::process_vary_header;
 use crate::services::subgraph;
 use crate::services::supergraph;
 use crate::spec::query::EXTENSIONS_VALUE_COMPLETION_KEY;
-use crate::synchronization::Mutex;
+use crate::sync::Mutex;
 use crate::test_harness::make_fake_batch;
 
 // Test Vary processing
